Implementado novos consumers para a criação de movimentos; Ajustes nas telas de entrada e saída; Criação do pedido de compra de forma automática a partir do ponto de ressuprimento.

+++ common/src/SupplyFlow.Common/Entities/Movimento.cs
@@ -4,7 +4,7 @@ public class Movimento : IEntity
 {
     public Guid Id { get; set; }
     public Produto Produto { get; set; }
-    public decimal Quantidade { get; set; }
+    public int Quantidade { get; set; }
     public TipoMovimento TipoMovimento { get; set; }
     public DateTime DataMovimento { get; set; }
 }
