+++ services/src/SupplyFlow.Movimento/MovimentoDto.cs
@@ -3,4 +3,4 @@ using SupplyFlow.Common.Entities;
 
 namespace SupplyFlow.Service.Dtos;
 
-public record CreateMovimentoDto([Required] Guid Produto, [Required] decimal Quantidade, [Required] TipoMovimento TipoMovimento);
+public record CreateMovimentoDto([Required] Guid Produto, [Required] int Quantidade, [Required] TipoMovimento TipoMovimento);
